wencai/util/requirements.txt
pillow
fitz
PyMuPDF
python-docx
pyDes
pdfminer
comtypes
moviepy --ignore-installed imageio
pytesseract
